block sessions on backend

+++ shortener/src/routes/api/update-link/+server.ts
@@ -1,11 +1,8 @@
+import { isPrivilegedSession } from '../../../hooks.server';
 import { getPayload } from '$lib/getPayload';
 import { error, type RequestHandler, json } from '@sveltejs/kit';
 import type { Link } from 'src/stores';
-
-export const INVALID_LINK_NAMES = new Set([
-  'auth',
-  'api',
-]);
+import { INVALID_LINK_NAMES } from '$lib/links';
 
 const getRedirectsKV = (platform: Readonly<App.Platform>) => {
   const REDIRECTS_KV = platform.env?.REDIRECTS_KV;
@@ -21,7 +18,11 @@ const getRedirectsKV = (platform: Readonly<App.Platform>) => {
   }
 }
 
-export const POST: RequestHandler = async ({ platform, request }) => {
+export const POST: RequestHandler = async ({ platform, request, locals }) => {
+  if (!await isPrivilegedSession(locals)) {
+    throw error(403, 'cannot create links without logging in.');
+  }
+
   const { put } = getRedirectsKV(platform);
 
   const { name, url, privileged } = await getPayload(request, { name: 'string', url: 'string', privileged: '?boolean' });
@@ -36,7 +37,11 @@ export const POST: RequestHandler = async ({ platform, request }) => {
   return json(`link with name "${linkName}" successfully created`, { status: 200 })
 }
 
-export const PUT: RequestHandler = async ({ platform, request }) => {
+export const PUT: RequestHandler = async ({ platform, request, locals }) => {
+  if (!await isPrivilegedSession(locals)) {
+    throw error(403, 'cannot edit links without logging in.');
+  }
+
   const { put, deleteItem } = getRedirectsKV(platform);
 
   const { name, url, privileged, oldName } = await getPayload(request, { name: 'string', oldName: 'string', url: 'string', privileged: '?boolean' });
@@ -51,7 +56,11 @@ export const PUT: RequestHandler = async ({ platform, request }) => {
   return json(`link with name "${linkName}" successfully updated`, { status: 200 })
 }
 
-export const DELETE: RequestHandler = async ({ platform, request }) => {
+export const DELETE: RequestHandler = async ({ platform, request, locals }) => {
+  if (!await isPrivilegedSession(locals)) {
+    throw error(403, 'cannot edit links without logging in.');
+  }
+
   const { deleteItem } = getRedirectsKV(platform);
 
   const { name } = await getPayload(request, { name: 'string' });
